project: early snapcraft.yaml validation
Avoid creating and invalid multipass instance by doing an early
validation of snapcraft.yaml. This has the benefit of faster turnaround
times too.

LP: #1794507
